requirements.txt
streamlit
transformers
torch
safetensors
accelerate
huggingface_hub
python-dotenv